client/requirements.txt
phe 
Flask-API 
flask_cors 
requests
pyopenssl
pycryptodome